RSDK-10108 RSDK-10315 follow module structure, Isolate C code (#49)
* isolate c code

* isolate c code

* Revert "isolate c code"

This reverts commit 00b98e7cefcdda5350be3fa1b00d952305dc0e66.

* edits

* lint

* tests

* fix test

* mock out c functions

* c style

* fixes and lint

* Delete .vscode/settings.json

* lint

* fix error

* make regions package

* remove function

* pr comments

+++ gateway/downlink_test.go
@@ -214,7 +214,7 @@ func TestCreateDownlink(t *testing.T) {
 				expectedMIC, err := crypto.ComputeLegacyDownlinkMIC(
 					types.AES128Key(tt.device.NwkSKey),
 					*types.MustDevAddr(tt.device.Addr),
-					initialFCntDown+1,
+					uint32(initialFCntDown)+1,
 					payloadWithoutMIC,
 				)
 				test.That(t, err, test.ShouldBeNil)
